# Flatten normalpath EqualsOrContainsPath check