This is a small optimization around the `EqualsOrContainsPath` method. It's called repeatedly over a filesystem when walking and using mappers or matchers. The check is now constant time for a given length. Added a micro benchmark to showcase on unix. Didn't adjust the windows method as theres some changes needed for volume handling.
```
                             │    sec/op     │    sec/op     vs base                 │
EqualsOrContainsPath/Short-8    45.36n ± ∞ ¹   15.56n ± ∞ ¹        ~ (p=1.000 n=1) ²
EqualsOrContainsPath/Long-8    458.20n ± ∞ ¹   15.56n ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                         144.2n         15.56n        -89.21%
```