@emcfarlane 2024-08-12
Can close, no real world diff as it's dominated by io.